+++ CHANGELOG.md
@@ -21,7 +21,7 @@ Types of changes
 ### Added
 
 - tasks explicitly request ephemeral storage
-- `custom_task` and `custom_memory_optimized_task` allow selecing storage size
+- `custom_task` and `custom_memory_optimized_task` allow selecting storage size
 
 
 ## 2.27.4 - 2023-07-18