+++ CHANGELOG.md
@@ -21,7 +21,8 @@ Types of changes
 ### Added
 
 - tasks explicitly request ephemeral storage
-- `custom_task` and `custom_memory_optimized_task` allow selecting storage size
+- `custom_task` allows selecting storage size
+- `custom_memory_optimized_task` removed -- functionality merged with `custom_task`
 
 
 ## 2.27.4 - 2023-07-18