+++ CHANGELOG.md
@@ -22,7 +22,11 @@ Types of changes
 
 - Tasks explicitly request ephemeral storage
 - `custom_task` and `custom_memory_optimized_task` allow selecting storage size
-- `custom_memory_optimized_task` removed -- functionality merged with `custom_task`
+- `custom_memory_optimized_task` functionality merged into `custom_task`
+
+### Deprecated
+
+- `custom_memory_optimized_task`
 
 ## 2.27.4 - 2023-07-18